feat(platform): re-structure with mock trpc endpoints
