+++ apps/platform/src/app/[locale]/auth/error/page.tsx
@@ -1,11 +1,7 @@
-import { TLocale } from "@maany_shr/e-class-translations";
-import ErrorPage from "../../../../components/login-error";
-import { getLocale } from "next-intl/server";
+import LoginErrorServerComponent from "apps/platform/src/lib/infrastructure/server/pages/auth/login-error-rsc";
 
 export default async function Login() {
-  const platform = process.env.E_CLASS_PLATFORM_SHORT_NAME || 'E-Class';
-  const locale = await getLocale();
   return (
-    <ErrorPage platform={platform} locale={locale as TLocale}/>
+    <LoginErrorServerComponent/>
   )
 }
