+++ apps/platform/src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import AuthContext from "../../../../auth/config"
+import AuthContext from "../../../../lib/infrastructure/server/config/auth/next-auth.config"
 
 const handlers = AuthContext.handlers
 export const { GET, POST } = handlers
