+++ apps/platform/src/lib/infrastructure/client/hooks/useSuspenseQuery.ts
@@ -1,0 +1,3 @@
+export function useSuspenseQuery<T>(input: T): T {
+    return input;
+}
