+++ apps/platform/src/lib/infrastructure/client/mocks/trpc-mocks.ts
@@ -1,0 +1,7 @@
+import { skillsQuery } from "./skills-mock";
+
+export const MockedTrpcClient = {
+    getSkills: {
+        queryOptions: skillsQuery,
+    }
+}
