+++ apps/platform/src/lib/infrastructure/server/config/auth/next-auth.config.ts
