+++ apps/platform/tsconfig.json
@@ -33,7 +33,7 @@
     ".next/types/**/*.ts",
     "postcss.config.mjs",
     "eslint.config.mjs"
-, "src/auth/config.ts", "../../packages/auth/src/infrastructure/gateway/next-auth-gateway.ts"  ],
+, "src/lib/infrastructure/server/config/auth/next-auth.config.ts", "../../packages/auth/src/infrastructure/gateway/next-auth-gateway.ts"  ],
   "exclude": [
     "node_modules",
     "jest.config.ts",
